Module Vcs_git_blocking

Implementation of a git provider for the Vcs library, based on Stdlib and Git_cli.

This implementation is based on the git command line tool. We run it as an external program with utils from Stdlib and Unix, producing the right command line invocation and parsing the output to produce a typed version of the expected results with Git_cli. Note that git must be found in the PATH of the running environment.

type 'a t = [> Git_cli.Trait.t ] as 'a Vcs.t

type t' = Git_cli.Trait.t t

This is a convenient type alias that may be used to designate a provider with the exact list of traits supported by this implementation.

val create : Base.unit -> _ t